requirements.txt
asgiref==3.4.1
assertpy==1.1
astroid==2.9.3
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
dodgy==0.2.1
Faker==12.0.0
flake8==3.9.2
flake8-polyfill==1.0.2
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pep8-naming==0.10.0
platformdirs==2.4.1
pluggy==1.0.0
prospector==1.6.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
pylint==2.12.2
pylint-celery==0.3
pylint-django==2.5.0
pylint-flask==0.6
pylint-plugin-utils==0.7
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requirements-detector==0.7
setoptconf-tmp==0.3.1
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
